修改缓存逻辑

+++ pojo/src/main/java/com/platon/browser/dto/block/BlockInfo.java
@@ -12,7 +12,7 @@ public class BlockInfo {
 
     private String node;
 
-    private int transaction;
+    private long transaction;
 
     private String blockReward;
 }
